First commit
Values can be successfully entered through an HTML form and sent to a JS file to create a "Substituents" object. 
Currently, there is no working data validation, so a user can enter anything and it will be accepted. As well, a user can leave fields blank.  Both of these should be fixed in subsequent versions.