Fixed validate function; dynamic forms work
validate function now works. As well, dynamic forms can now be generated based off of the number of monomers entered for each functional group. If user entered data is not acceptable (e.g. 'text' for a number variable like num), then the form cannot be generated until an acceptable input is entered.
Issues: a user can click the submit button on the first form and generate an indefinite number of forms. As well, entering data into the dynamic form does nothing as of this moment.