Changed Data Type Validation
checkDataTypes function is now more flexible in its design, where its methods are not hardcoded. Instead, it checks values within a class (int and float classes) and returns 'false' if any of them are considered invalid. At this moment, however, the string method is still hardcoded because no other string variables are expected to be accepted elsewhere.
The function has yet to be implemented for the dynamic form group, but it will be in subsequent versions. As such, the i variable used to loop will have to be changed so that it is the first number in the list of group items.