Improved data sorting
Dynamically generated input fields now accept a blank string as an input and interprets it as an unknown value (0).
A reference group can now be found and it will choose the better of the two groups based on what is known (e.g. if it has more mass values or more known values).  Will need to change in the future so that functional groups with more monomers are not always chosen since they will almost always have more knowns than those with fewer monomers.
A calculation route can now be determined for the reference group through routeFinder(), however the function is unfinished and the complimentary routes have not yet been implemented.