Alpha Version 1.0 (a1.0)
All calculation routes have been implemented for both reference and complimentary groups.  Debugging will be necessary and there are likely numerous optimizations that can be made (e.g. certain actions that are repeated across multiple calculation routes can likely be made into a re-usable function).
Calculation results are now displayed on the webpage via the newly added element_generator. For the future, 'form_generator' should probably be moved to this file.
Fixed an issue with how complimentary groups are assigned in data_input_entry.

This version represents the first, finished prototype of the web application. In its current state, it has worked as intended for all situations tested, and the functionality of the original C++ program has been met and, in fact, surpassed by this application.  As such, this version will be referred to as 'Alpha Version 1.0 (a1.0)', and all subsequent versions will be a1.1, a1.2, etc. If the three features outlined in the READ ME file are all implemented, then the application will have gone past Alpha, and entered Beta.