Overhauled data sorting, calculations, display
Data sorting has been simplified and some calculation routes were changed as a result. 
Invalid percent values are rejected (percent sums greater than 100, less than 100 if all are given, etc).
If there is only 1 comonomer in a group, then the percent field is disabled and a placeholder for '100' is displayed.
Minor changes/optimizations.