Improved reference calculations + misc changes
The "Percent and Mass" route has been changed to the excess information route for the reference group. This desicion was made to create a calculation route that would handle excess information given by the user with the ability to detect possible source of error.
Fixed an issue where changing the number of comonomers from 2 to 1 would cause issues when entering new values for percent.